backend/requirements.txt
flask
flask-cors
groq
python-dotenv
requests
beautifulsoup4
sentence-transformers
scikit-learn